requirements.txt
sqlalchemy==2.0.23
pymysql==1.1.0
faker==20.1.0
python-dotenv==1.0.0
cryptography==41.0.7
